backend/requirements.txt
fastapi
uvicorn
python-multipart

torch==2.1.2
torchaudio==2.1.2

pyannote.audio==3.1.1
pydub
ffmpeg-python
